fastapi/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
PyMySQL
python-dotenv
cryptography